T08
- 게임 매니저에서 새로운 게임 등록
- 게임 데이터 String-Array간 변환하는 로직 Board로 이동
